Harmonise l'import des mappings publicodes
